Check UDIO response length
The "read states" response on my Smart Lock 3.0 has 80 bytes and is sent
as 4 packets with 20 bytes each - exactly the MTU. It never got through
(resulting in a timeout) because the previous code was waiting for a
less-than-MTU-sized packet before trying to decrypt it.

With this change the packet length encoded in the encrypted command is
used to check whether the response received so far is complete.